requirements.txt
Django==5.0.4
djangorestframework==3.15.1
beautifulsoup4
openai
selenium
django-filter